nuevos estilos

+++ src/components/DisplayProvedor.vue
@@ -12,13 +12,7 @@
       </div>
       <div class="column is-8">
       </div>
-      <div class="column is-1">
-        <router-link :to="{ name: 'Init'}" class="button is-link is-rounded">
-          <span class="icon is-small is-left">
-            <i class="fas fa-sign-out-alt fa-lg"></i>
-          </span>
-        </router-link>
-      </div>
+
     </div>
     <div class="columns">
       <div class="column is-10 is-offset-1">
@@ -41,11 +35,11 @@
               <td><a v-bind:href="'mailto:'+provedor.email">{{provedor.email}}</a></td>
               <td>{{provedor.description}}</td>
               <td>
-                <a href="#" v-on:click="deleteProvedor(provedor.id, provedor.name)">
+                <span v-on:click="deleteProvedor(provedor.id, provedor.name)">
                   <span class="icon is-small is-left">
                     <i class="fas fa-trash-alt fa-lg"></i>
                   </span>
-                </a>
+                </span>
                 <router-link :to="{ name: 'UpdateProvedor', params: {id: provedor.id} }">
                   <span class="icon is-small is-left">
                     <i class="fas fa-pen-square fa-lg"></i>
@@ -55,20 +49,6 @@
             </tr>
           </tbody>
         </table>
-      </div>
-    </div>
-    <div class="columns">
-      <div class="column is-2 is-offset-10">
-        <p class="control">
-          <a class="button is-light">
-            <router-link :to="{ name: 'MenuAlmacenista'}" class="button is-link is-rounded">
-              <span class="icon is-small is-left">
-                <i class="fas fa-arrow-left fa-lg"></i>
-              </span>
-              . Volver al menú
-            </router-link>
-          </a>
-        </p>
       </div>
     </div>
   </div>
@@ -94,19 +74,26 @@
         .catch(err => console.log(err));
       },
       deleteProvedor(id,name){
-        const response = confirm('Estas seguro de eliminar al provedor '+name+'?');
+        const response = confirm('Estas seguro de eliminar los datos del proveedor '+name+'?');
         if(response){
-          this.axios.delete(url+"provedor/"+id)
-          .then(res =>{
-            alert(res.data.provedor);
-            for(var i=0; i<this.provedors.length; i++){
-              if(this.provedors[i].id==id){
-                this.provedors.splice(i,1);
-                break
-              }
+          this.axios.get(url+"product/provedor/"+id).then(res=>{
+            if(res.data.length<=0){
+              this.axios.delete(url+"provedor/"+id)
+              .then(res =>{
+
+                alert(res.data.provedor);
+                for(var i=0; i<this.provedors.length; i++){
+                  if(this.provedors[i].id==id){
+                    this.provedors.splice(i,1);
+                    break
+                  }
+                }
+              })
+              .catch(err => console.log(err));
+            }else{
+             alert("No se han podido eliminar los datos del proveedor "+name+" debido a que tiene productos asociados!!");
             }
           })
-          .catch(err => console.log(err));
         }
       }
     }
